Expose the new macro to users
