+++ Sources/DependenciesMacros/DependenciesMacros.swift
@@ -1,11 +1,2 @@
-// The Swift Programming Language
-// https://docs.swift.org/swift-book
-
-/// A macro that produces both a value and a string containing the
-/// source code that generated the value. For example,
-///
-///     #stringify(x + y)
-///
-/// produces a tuple `(x + y, "x + y")`.
-@freestanding(expression)
-public macro stringify<T>(_ value: T) -> (T, String) = #externalMacro(module: "DependenciesMacrosMacros", type: "StringifyMacro")
+@attached(member, names: named(encode), named(init), named(SubType), named(CodingKeys))
+public macro EnumCodable() = #externalMacro(module: "DependenciesMacrosMacros", type: "EnumCodableMacro")
